add a test for dark mode

+++ knightassist_mobile_app/lib/src/features/students/data/students_repository.dart
@@ -255,8 +255,8 @@ class StudentsRepository {
       String? lastName,
       String? email,
       String? profilePicPath,
-      int? totalVolunteerHours,
-      int? semesterVolunteerHourGoal,
+      double? totalVolunteerHours,
+      double? semesterVolunteerHourGoal,
       List<String>? categoryTags) async {
     Map<String, dynamic> params = {'studentID': studentID};
     if (firstName != null) params['firstName'] = firstName;
